Move all go code into the `proxy-init` directory (#96)
In preparation for introducing components other than the `proxy-init`
binary, this change moves all of the current Go code into a
`./proxy-init` directory:

* Move `iptables` to `proxy-init/internal/iptables`
* Move `ports` to `proxy-init/internal/util`
* Rename `IsValid` to `IsValidPort`

This changes the relevant Just recipes to be prefixed with
`proxy-init-`.
