+++ .github/workflows/go.yml
@@ -20,12 +20,12 @@ jobs:
     container: ghcr.io/linkerd/dev:v22-go
     steps:
       - uses: actions/checkout@2541b1294d2704b0964813337f33b291d3f8596b
-      - run: just fmt
+      - run: just go-fmt-check
 
   unit-test:
     runs-on: ubuntu-20.04
     container: ghcr.io/linkerd/dev:v22-go
     steps:
       - uses: actions/checkout@2541b1294d2704b0964813337f33b291d3f8596b
-      - run: just test-unit
+      - run: just proxy-init-test-unit
