+++ proxy-init/internal/util/portrange_test.go
@@ -1,4 +1,4 @@
-package ports
+package util
 
 import (
 	"reflect"
